refactor: split app into focused modules
